Created a Component attribute with a custom Editor.
The Component attribute can be applied to MonoBehaviour scripts. this attribute allows you to give the script a title and a description, which will show up in the editor at the top of the script.

If a MonoBehaviour script does not have this attribute, BaseEditor will automatically give that script a title based on it class name.

This functionallity only applies to scripts with the same namespace as the root namespace set in the project settings.